- moved MavenModuleSet.Builder to new top-level class Maven2Builder - removed duplicated FilterImpl classes
